x86/sev: Add snp_msg_seqno() helper
The SNP guest request message header contains a message count. The
message count is used while building the IV. The PSP firmware increments
the message count by 1, and expects that next message will be using the
incremented count. The snp_msg_seqno() helper will be used by driver to
get the message sequence counter used in the request message header,
and it will be automatically incremented after the request is successful.
The incremented value is saved in the secrets page so that the kexec'ed
kernel knows from where to begin.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>