x86/mm: Reserve unaccepted memory bitmap
A given page of memory can only be accepted once.  The kernel has a need
to accept memory both in the early decompression stage and during normal
runtime.

Use a bitmap to communicate the acceptance state of each page between
the decompression stage and normal runtime.  This eliminates the
possibility of attempting to double-accept a page.

Allocate the bitmap during decompression stage and hand it over to the
main kernel image via boot_params.

In the runtime kernel, reserve the bitmap's memory to ensure nothing
overwrites it.

Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Acked-by: Mike Rapoport <rppt@linux.ibm.com>