KVM: SVM: Restore setting of virtual interrupt bit
When running SNP in a mixed VMPL environment and one VMPL level is
running with restricted injection support, the setting of the virtual
interrupt (V_IRQ) bit in the VMCB can be lost. This can result in a vCPU
not being able to schedule an IPI, which can result in a guest hang.

If running as an SNP guest, restore the virtual interrupt bit at VMEXIT
if the V_INTR_PRIO value is non-zero.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>