x86/sev: Perform PVALIDATE using the SVSM_CORE_PVALIDATE call
The PVALIDATE instruction can only be performed at VMPL0. An SVSM will
be present when running at VMPL1 or lower.

When an SVSM is present, use the SVSM_CORE_PVALIDATE call to perform
memory validation instead of issuing the PVALIDATE instruction directly.

The pvalidate_page() function is now moved into sev-common.c and renamed
to early_pvalidate_page() so that it can be used by both the decompressor
code and early kernel code.

Signed-off-by: Tom Lendacky <thomas.lendacky@amd.com>